requirements.txt
fastapi
pydantic
typing
mysql-connector-python  # for MySQL
uvicorn  # for running the FastAPI application